requirements.txt
psycopg2-binary>=2.9.5
pandas>=1.5.0
numpy>=1.21.0
tqdm>=4.64.0
matplotlib>=3.5.0